requirements.txt
pyTelegramBotAPI
openai
python-dotenv
google-generativeai
pymongo